Refactor hamiltonian and models to reorganize spectrum functions and add Kitaev chain calculations
